pyproject.toml
[tool.poetry]
name = "district-selection"
version = "0.1.0"
description = "Way for me to pick a district"
authors = ["Adrian Obleton <obletonadrian@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
openai = "^1.40.3"
firecrawl-py = "^0.0.20"
backoff = "^2.2.1"
google-generativeai = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
